Create NeuralNetMagic
NeuralNetMagic is a Python toolkit for building and training simple neural networks using TensorFlow. With NeuralNetMagic, you can easily construct neural network architectures, train them on your data, and evaluate their performance. This toolkit provides a user-friendly interface to create neural networks with customizable input shapes and numbers of classes. It includes functionalities for defining the network structure, training the model, and evaluating its accuracy. Whether you're a beginner exploring the basics of neural networks or an experienced practitioner looking for a quick prototyping tool, NeuralNetMagic simplifies the process of neural network development and experimentation. Unlock the power of neural networks with NeuralNetMagic and dive into the world of deep learning.